+++ Packages/ai.natml.vision.blazepose/Runtime/BlazePosePipeline.cs
@@ -1,12 +1,13 @@
 /* 
 *   BlazePose
-*   Copyright (c) 2022 NatML Inc. All Rights Reserved.
+*   Copyright (c) 2023 NatML Inc. All Rights Reserved.
 */
 
 namespace NatML.Vision {
 
     using System;
     using System.Collections.Generic;
+    using System.Threading.Tasks;
     using UnityEngine;
     using NatML.Features;
     using NatML.Internal;
@@ -21,22 +22,6 @@ namespace NatML.Vision {
 
         #region --Client API--
         /// <summary>
-        /// Create the BlazePose pipeline.
-        /// </summary>
-        /// <param name="detector">BlazePose detector model data.</param>
-        /// <param name="predictor">BlazePose landmark predictor model data.</param>
-        /// <param name="maxDetections">Maximum number of detections to return.</param>
-        public BlazePosePipeline (MLModelData detector, MLModelData predictor, int maxDetections = Int32.MaxValue) {
-            this.detectorData = detector;
-            this.predictorData = predictor;
-            this.detectorModel = new MLEdgeModel(detector);
-            this.predictorModel = new MLEdgeModel(predictor);
-            this.detector = new BlazePoseDetector(detectorModel);
-            this.predictor = new BlazePosePredictor(predictorModel);
-            this.maxDetections = maxDetections;
-        }
-
-        /// <summary>
         /// Detect poses in an image.
         /// </summary>
         /// <param name="inputs">Input image. This MUST be an `MLImageFeature`.</param>
@@ -50,8 +35,6 @@ namespace NatML.Vision {
             if (input == null)
                 throw new ArgumentException(@"BlazePose pipeline expects an image feature", nameof(inputs));
             // Detect poses
-            (input.mean, input.std) = detectorData.normalization;
-            input.aspectMode = detectorData.aspectMode;
             var detectedPoses = detector.Predict(input);
             // Predict landmarks
             var capacity = Mathf.Min(detectedPoses.Length, maxDetections);
@@ -59,13 +42,13 @@ namespace NatML.Vision {
             for (var i = 0; i < capacity; ++i) {
                 // Extract ROI
                 var detection = detectedPoses[i];
-                var roi = input.RegionOfInterest(detection.regionOfInterest, -detection.rotation, Color.black);
+                var detectionRect = detection.regionOfInterest;
+                var roi = new MLImageFeature(detectionRect.width, detectionRect.height);
+                input.CopyTo(roi, detectionRect, -detection.rotation, Color.black);
                 // Predict landmarks
-                (roi.mean, roi.std) = predictorData.normalization;
-                roi.aspectMode = predictorData.aspectMode;
                 var landmarks = predictor.Predict(roi);
                 // Create pose
-                var inputType = predictorModel.inputs[0] as MLImageType;
+                var inputType = predictor.model.inputs[0] as MLImageType;
                 var keypoints = new BlazePosePredictor.Keypoints(landmarks.keypoints.data, inputType, detection.regionOfInterestToImageMatrix);
                 var pose = new BlazePosePredictor.Pose(landmarks.score, keypoints, landmarks.keypoints3D);
                 result.Add(pose);
@@ -78,20 +61,40 @@ namespace NatML.Vision {
         /// Dispose the pipeline and release resources.
         /// </summary>
         public void Dispose () {
-            detectorModel.Dispose();
-            predictorModel.Dispose();
+            detector.Dispose();
+            predictor.Dispose();
+        }
+
+        /// <summary>
+        /// Create the BlazePose pipeline.
+        /// </summary>
+        /// <param name="maxDetections">Maximum number of detections to return.</param>
+        /// <param name="detector">BlazePose detector model data.</param>
+        /// <param name="predictor">BlazePose landmark predictor model data.</param>
+        /// <returns>BlazePose pipeline.</returns>
+        public static async Task<BlazePosePipeline> Create (
+            int maxDetections = Int32.MaxValue,
+            BlazePoseDetector detector = null,
+            BlazePosePredictor predictor = null
+        ) {
+            detector ??= await BlazePoseDetector.Create();
+            predictor ??= await BlazePosePredictor.Create();
+            var pipeline = new BlazePosePipeline(detector, predictor, maxDetections);
+            return pipeline;
         }
         #endregion
 
 
         #region --Operations--
-        private readonly MLModelData detectorData;
-        private readonly MLModelData predictorData;
-        private readonly MLEdgeModel detectorModel;
-        private readonly MLEdgeModel predictorModel;
         private readonly BlazePoseDetector detector;
         private readonly BlazePosePredictor predictor;
         private readonly int maxDetections;
+
+        private BlazePosePipeline (BlazePoseDetector detector, BlazePosePredictor predictor, int maxDetections) {
+            this.detector = detector;
+            this.predictor = predictor;
+            this.maxDetections = maxDetections;
+        }
         #endregion
     }
 }