+++ Packages/ai.natml.vision.blazepose/Changelog.md
@@ -1,3 +1,10 @@
+## 1.0.2
++ Added `BlazePosePipeline.Create` static method for creating the pipeline.
++ Added `BlazePosePredictor.Create` static method for creating the predictor.
++ Removed `BlazePosePipeline` constructor. Use `Create` method instead.
++ Removed `BlazePosePredictor` constructor. Use `Create` method instead.
++ Upgraded to NatML 1.1.16.
+
 ## 1.0.1
 + Updated `README`.